+++ packages/server/gateway/views/layoutFramed.hjs
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <!-- Required meta tags -->
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
+    {{$styles}}{{/styles}}
+
+    <title>{{ title }}</title>
+</head>
+<body>
+    <nav class="navbar navbar-dark" style="background: #004B8B">
+        <a class="navbar-brand" href="/">
+            <img src="/public/images/bindy.svg" style="vertical-align: text-bottom; height: 26px;">
+            Fluid
+        </a>
+        <div id="connection-state" style="color: #fff"></div>
+        <div id="fluid-user">
+        </div>
+    </nav>
+
+    <!-- Scripts -->
+    <script>
+        var user = {{{ user }}};
+    </script>
+    <script src="{{#hfurl}}/public/scripts/dist/loaderHost.js{{/hfurl}}"></script>
+    {{$scripts}}{{/scripts}}
+    {{!-- Content After Scripts so that Outer Document Service can start fetching data --}}
+    {{$content}}{{/content}}
+
+    <script>
+        loaderHost.navbar.initialize(user);
+    </script>
+</body>
+</html>
