+++ packages/server/gateway/views/loaderHost.hjs
@@ -1,0 +1,42 @@
+{{<layoutFramed}}
+
+{{$content}}
+    <iframe onload="this.style.opacity=1;"
+            style="opacity:0;border:none;width:100%;height:100%;"
+            id="ifr"
+            sandbox="allow-scripts"
+            srcdoc='{{> loaderFramed}}'>
+    </iframe>
+{{/content}}
+
+{{$scripts}}
+
+    <script>
+        var cache = {{{ cache }}};
+        var timings = "{{ timings }}";
+        {{#chaincode}}
+        var chaincode = {{{ chaincode }}};
+        {{/chaincode}}
+        {{^chaincode}}
+        var chaincode;
+        {{/chaincode}}
+        var url = "{{ url }}";
+        var resolved = {{{ resolved }}};
+        var jwt = "{{ jwt }}";
+        var npm = "{{ npm }}";
+        var config = {{{ config }}};
+        var clientId = "{{ clientId }}";
+
+        var scriptIds = [];
+  
+        loaderHost.initialize(
+            window.location.href,
+            resolved,
+            cache,
+            jwt,
+            config,
+            clientId);
+    </script>
+{{/scripts}}
+
+{{/layoutFramed}}
